docker/bridge/requirements.txt
influxdb==5.3.0
paho_mqtt==1.5.0
